requirements.txt
# Web framework
fastapi==0.104.1
starlette==0.27.0
uvicorn[standard]==0.24.0

# File handling
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0

# Data processing
pandas>=1.3.0
python-pptx>=0.6.21
python-docx>=0.8.11

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Utilities
requests>=2.26.0
tqdm>=4.62.0
Pillow>=9.0.0
python-magic>=0.4.27
nltk>=3.8
cryptography>=41.0.0
prometheus-client>=0.17.0
fonttools>=4.38.0
httpx==0.24.1
